requirements.txt
aniso8601==9.0.1
anyio==3.7.1
asgiref==3.7.2
click==8.1.6
Django==4.2.3
django-admin-display==1.3.0
django-debug-toolbar==4.1.0
django-guardian==2.4.0
exceptiongroup==1.1.2
graphene==3.2.2
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.4
libcst==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.1
Pillow==10.0.0
promise==2.3
Pygments==2.15.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.4.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.30.0
strawberry-django-jwt==0.2.3
strawberry-graphql==0.195.2
strawberry-graphql-django==0.14.0
text-unidecode==1.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uvicorn==0.21.1
